// src/utils/types.ts
export type NasaPictureOfDayProps = {
  date: string;
  explanation: string;
  hdurl: string;
  media_type: string;
  title: string;
  url: string;
}